fix bugs

+++ android/src/main/java/com/mazadatimagepicker/Camera/PickerCameraActivity.java
@@ -395,7 +395,7 @@ public class PickerCameraActivity extends AppCompatActivity {
           RectF rectF = imageItems.get(i).getZoomImage().getZoomedRect();
           runOnUiThread(() -> loadingCl.setVisibility(View.VISIBLE));
           Bitmap bitmap = BitmapFactory.decodeFile(imageItems.get(i).getFile().getPath());
-          if(bitmap.getWidth() > 0 && (bitmap.getWidth() * rectF.width()) > 0) {
+          if(bitmap.getWidth() > 0 && ((int)(bitmap.getWidth() * rectF.width())) > 0) {
             bitmap = Bitmap.createBitmap(bitmap, (int) (bitmap.getWidth() * rectF.left), (int) (bitmap.getHeight() * rectF.top), (int) (bitmap.getWidth() * rectF.width()), (int) (bitmap.getHeight() * rectF.height()));
             File file = ImageUtils.bitmapToFile(PickerCameraActivity.this, bitmap, imageItems.get(i).getPercentage());
             imageItems.get(i).setFile(file);
